Refactor SurfaceKind in the GUI and move surface implementations
