+++ crates/cherry-rs/src/core/surfaces/kinds/mod.rs
@@ -1,0 +1,13 @@
+pub mod conic;
+pub mod image;
+pub mod iris;
+pub mod object;
+pub mod probe;
+pub mod sphere;
+
+pub use conic::Conic;
+pub use image::Image;
+pub use iris::Iris;
+pub use object::Object;
+pub use probe::Probe;
+pub use sphere::Sphere;
